Break a source code dependency cycle
