+++ src/PluginInterface.h
@@ -42,15 +42,43 @@
 #ifndef __AUDACITY_PLUGININTERFACE_H__
 #define __AUDACITY_PLUGININTERFACE_H__
 
-#include "ConfigInterface.h"
 #include "EffectInterface.h"
 #include "ComponentInterface.h"
 #include "Identifier.h"
 #include "ModuleInterface.h"
+#include <variant>
 
 class ModuleInterface;
 
-namespace PluginSettings { enum ConfigurationType : unsigned; }
+namespace PluginSettings {
+
+enum ConfigurationType : unsigned {
+   Shared, Private
+};
+
+//! Supported types for settings
+using ConfigValueTypes = std::tuple<
+     wxString
+   , int
+   , bool
+   , float
+   , double
+>;
+
+//! Define a reference to a variable of one of the types in ConfigValueTypes
+/*! Avoid repetition of the list of types */
+template<bool is_const, typename> struct ConfigReferenceGenerator;
+template<bool is_const, typename... Types>
+struct ConfigReferenceGenerator<is_const, std::tuple<Types...>> {
+   using type = std::variant< std::reference_wrapper<
+      std::conditional_t<is_const, const Types, Types> >... >;
+};
+using ConfigReference =
+   ConfigReferenceGenerator<false, ConfigValueTypes>::type;
+using ConfigConstReference =
+   ConfigReferenceGenerator<true, ConfigValueTypes>::type;
+
+}
 
 class AUDACITY_DLL_API PluginManagerInterface /* not final */
 {
